Break stuff with major changes
Apologies for this; in an effort to keep all per-agent promise
configuration in a single place (/usr/local/etc/cfengine.json),
this commit has breaking changes. In particular, the old system
for backups to S3 has been modified, and manual changes are
needed to play nice with the new setup. Will try to not have
a breaking change like this again for S3 backups.

In addition, promises have been cleaned up and documentation
has been improved.